+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 - Deprecated - cli.parseValue() overload w/o arg source
 - Added - opt.initConfig() customization point
 - Added - Transform actions to modify source arg before parsing
+- Added - Allow args from more than just one environment variable
 
 ## dimcli 7.5.0 (2025-11-20)
 - Added - cli.success() and cli.fail() overloads with just the exit code